Refactor Geogebra Admin View, Geogebra Interface and create BaseGeogebraClient
